+++ components/reusable/navbar.tsx
@@ -5,7 +5,7 @@ import { type Locale } from '@/i18n-config';
 import LocaleSwitcher from '@/components/reusable/locale-switcher';
 import { ModeToggle } from '@/components/reusable/theme-toggle';
 import { Button } from '@/components/ui/button';
-import { useSyncExternalStore } from 'react';
+import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { AuthService } from '@/lib/requests';
 import { deleteCookie, getCookie } from 'cookies-next';
@@ -23,36 +23,24 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from '@/components/ui/tooltip';
-import { parseSessionUserCookie } from '@/lib/auth-role';
 import { Bot } from 'lucide-react';
 import { type Dictionary } from '@/get-dictionary';
 
 type NavbarUser = { userId: string; isAdmin: boolean };
-
-const noop = () => {};
-
-let hydrated = false;
-
-const subscribeHydration = (onStoreChange: () => void) => {
-  if (!hydrated) {
-    hydrated = true;
-    onStoreChange();
-  }
-
-  return noop;
-};
-
-const getHydrationSnapshot = () => hydrated;
 
 const readUserFromCookies = (): NavbarUser | undefined => {
   const userCookie = getCookie('user');
   const tokenCookie = getCookie('token');
   if (!userCookie || !tokenCookie) return undefined;
 
-  const parsed = parseSessionUserCookie(userCookie as string);
-  if (!parsed) return undefined;
+  try {
+    const parsed = JSON.parse(userCookie as string) as NavbarUser;
+    if (parsed && typeof parsed.userId === 'string') return parsed;
+  } catch {
+    return undefined;
+  }
 
-  return { userId: parsed.userId, isAdmin: parsed.isAdmin };
+  return undefined;
 };
 
 export default function Navbar({
@@ -63,12 +51,7 @@ export default function Navbar({
   dictionary: Dictionary;
 }) {
   const router = useRouter();
-  const isHydrated = useSyncExternalStore(
-    subscribeHydration,
-    getHydrationSnapshot,
-    () => false
-  );
-  const user = isHydrated ? readUserFromCookies() : undefined;
+  const [user, setUser] = useState<NavbarUser | undefined>(readUserFromCookies);
 
   const handleLogout = async () => {
     try {
@@ -84,6 +67,7 @@ export default function Navbar({
     } catch {}
     deleteCookie('token');
     deleteCookie('user');
+    setUser(undefined);
     router.push(`/${lang}/login`);
   };
 
@@ -216,83 +200,81 @@ export default function Navbar({
 
         <div className="flex items-center gap-2 md:gap-3">
           <div className="flex items-center gap-2 md:gap-3">
-            {isHydrated ? (
-              user ? (
-                <>
+            {user ? (
+              <>
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <Link
+                      href={`/${lang}/profile`}
+                      className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
+                    >
+                      {dictionary.pages.home.navigation.profile}
+                    </Link>
+                  </TooltipTrigger>
+                  <TooltipContent>
+                    <p>{dictionary.tooltips.profile}</p>
+                  </TooltipContent>
+                </Tooltip>
+                {user.isAdmin && (
                   <Tooltip>
                     <TooltipTrigger asChild>
                       <Link
-                        href={`/${lang}/profile`}
+                        href={`/${lang}/admin`}
                         className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
                       >
-                        {dictionary.pages.home.navigation.profile}
+                        {dictionary.pages.home.navigation.adminDashboard}
                       </Link>
                     </TooltipTrigger>
                     <TooltipContent>
-                      <p>{dictionary.tooltips.profile}</p>
+                      <p>{dictionary.tooltips.adminDashboard}</p>
                     </TooltipContent>
                   </Tooltip>
-                  {user.isAdmin && (
-                    <Tooltip>
-                      <TooltipTrigger asChild>
-                        <Link
-                          href={`/${lang}/admin`}
-                          className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
-                        >
-                          {dictionary.pages.home.navigation.adminDashboard}
-                        </Link>
-                      </TooltipTrigger>
-                      <TooltipContent>
-                        <p>{dictionary.tooltips.adminDashboard}</p>
-                      </TooltipContent>
-                    </Tooltip>
-                  )}
-                  <Tooltip>
-                    <TooltipTrigger asChild>
-                      <Button
-                        size="sm"
-                        className="h-9 px-4"
-                        onClick={handleLogout}
-                      >
-                        {dictionary.pages.home.navigation.logout}
-                      </Button>
-                    </TooltipTrigger>
-                    <TooltipContent>
-                      <p>{dictionary.tooltips.logout}</p>
-                    </TooltipContent>
-                  </Tooltip>
-                </>
-              ) : (
-                <>
-                  <Tooltip>
-                    <TooltipTrigger asChild>
-                      <Link
-                        href={`/${lang}/login`}
-                        className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
-                      >
-                        {dictionary.pages.home.navigation.login}
-                      </Link>
-                    </TooltipTrigger>
-                    <TooltipContent>
-                      <p>{dictionary.tooltips.signIn}</p>
-                    </TooltipContent>
-                  </Tooltip>
-                  <Tooltip>
-                    <TooltipTrigger asChild>
-                      <Link
-                        href={`/${lang}/register`}
-                        className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
-                      >
-                        {dictionary.pages.home.navigation.register}
-                      </Link>
-                    </TooltipTrigger>
-                    <TooltipContent>
-                      <p>{dictionary.tooltips.createAccount}</p>
-                    </TooltipContent>
-                  </Tooltip>
-                </>
-              )
-            ) : undefined}
+                )}
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <Button
+                      size="sm"
+                      className="h-9 px-4"
+                      onClick={handleLogout}
+                    >
+                      {dictionary.pages.home.navigation.logout}
+                    </Button>
+                  </TooltipTrigger>
+                  <TooltipContent>
+                    <p>{dictionary.tooltips.logout}</p>
+                  </TooltipContent>
+                </Tooltip>
+              </>
+            ) : user ? undefined : (
+              <>
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <Link
+                      href={`/${lang}/login`}
+                      className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
+                    >
+                      {dictionary.pages.home.navigation.login}
+                    </Link>
+                  </TooltipTrigger>
+                  <TooltipContent>
+                    <p>{dictionary.tooltips.signIn}</p>
+                  </TooltipContent>
+                </Tooltip>
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <Link
+                      href={`/${lang}/register`}
+                      className="text-muted-foreground hover:text-primary text-sm font-medium transition-colors"
+                    >
+                      {dictionary.pages.home.navigation.register}
+                    </Link>
+                  </TooltipTrigger>
+                  <TooltipContent>
+                    <p>{dictionary.tooltips.createAccount}</p>
+                  </TooltipContent>
+                </Tooltip>
+              </>
+            )}
           </div>
           <div className="flex items-center gap-2 md:gap-3">
             <Tooltip>
@@ -311,7 +293,7 @@ export default function Navbar({
                 <p>{dictionary.tooltips.toggleTheme}</p>
               </TooltipContent>
             </Tooltip>
-            {isHydrated && user === undefined ? (
+            {!user && (
               <Tooltip>
                 <TooltipTrigger asChild>
                   <Button size="sm" className="h-9 px-4">
@@ -322,7 +304,7 @@ export default function Navbar({
                   <p>{dictionary.tooltips.getStarted}</p>
                 </TooltipContent>
               </Tooltip>
-            ) : undefined}
+            )}
           </div>
         </div>
       </div>
