+++ tsconfig.json
@@ -19,8 +19,7 @@
       }
     ],
     "paths": {
-      "@/*": ["./*"],
-      "lucide-react": ["./lib/lucide-react"]
+      "@/*": ["./*"]
     }
   },
   "include": [
